🎨 Rename module: fasthttp/src -> fasthttp/app
This change sets the default executable name to "app".
We may also drop the "-o app" in the Dockerfile.
